feat: implement weapon 'magic wand'

+++ Assets/Scripts/common/stats/ExcelWeapon.cs
@@ -10,6 +10,6 @@ public class ExcelWeapon
     public int CriticalDamage;
     public string ProjectileName;
     public int ProjectileSize;
-    public float ProjectileSpeed;
+    public int ProjectileSpeed;
     public int ProjectileCount;
 }
